+++ comics/comics/hipsterhitler.py
@@ -1,4 +1,4 @@
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import CrawlerBase
 from comics.core.comic_data import ComicDataBase
 
 
@@ -7,19 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.hipsterhitler.com/'
     start_date = '2010-08-01'
+    end_date = '2013-01-15'
+    active = False
     rights = 'JC & APK'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 120
-    schedule = None
-    time_zone = 'US/Central'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://hipsterhitler.com/feed/')
-        for entry in feed.for_date(pub_date):
-            if 'Comics' not in entry.tags:
-                continue
-            url = entry.content0.src('img')
-            title = entry.title
-            return CrawlerImage(url, title)
+        pass  # Comic no longer published